Eliminate same disk check; always fall back to copy
We know the reality of file systems is that when you want to "move" a file within a single file system, it's just a matter of updating the directories so that it appears elsewhere in the tree, but the contents of the file on disk do not change at all; whereas, to "move" a file across file systems is not technically possible, but can be simulated by making an exact copy on the other system, and then deleting the original.

Our program is prepared to do that; it can call java.nio.file.Files.move, or it can execute a copy-and-delete method that we borrowed from the jEdit library.

Prior to this commit, though, we tried to examine the file systems to determine beforehand if the move crossed filesystem boundaries or not.  That process was error-prone and slow.

So, just don't worry about it.  Try to do a simple "move".  If it cannot be done, Java will tell us.  We'll get an exception, which we can catch, and then we can try copy-and-delete.

In fact, in my testing so far, it looks like java.nio.file.Files.move actually does "move" files across filesystems, at least on Windows.  (Under the hood, it is likely completely different implementations on different types of filesystems; in fact, even java.nio.file.Path is just an Interface.)

I'll try test on a few other types of filesystems after this is committed.